changed mag readings int to float

+++ telemetry_source/lib/sensors/src/sensorManager.h
@@ -30,9 +30,9 @@ private:
         int ACCEL_R = 0;
         int ACCEL_P = 0;
         int ACCEL_Y = 0;
-        int MAG_R = 0;
-        int MAG_P = 0;
-        int MAG_Y = 0;
+        float MAG_R = 0;
+        float MAG_P = 0;
+        float MAG_Y = 0;
         int AUTO_GYRO_ROTATION_RATE = 0;
         char GPS_TIME[DATA_SIZE] = "";
         float GPS_ALTITUDE = 0.0;
